requirements.txt
pillow
imageio-ffmpeg
imagehash
pytesseract
